refactor: move vcenter connect task runner logic to inspect task runner
Relates to JIRA: DISCOVERY-784
